requirements.txt
epdb
python-Levenshtein
textblob
textblob-aptagger
wordgrapher
irclog2html
python-dateutil
argparse
